Changed to uint

+++ h5pp/include/h5pp/details/h5ppDatasetProperties.h
@@ -17,7 +17,7 @@ namespace h5pp {
         int                  ndims;
         std::vector<hsize_t> dims;
         std::vector<hsize_t> chunkDims;
-        size_t               compressionLevel = 0;
+        unsigned int         compressionLevel = 0;
 
         ~DatasetProperties() {
             H5Tclose(dataType);
